pyproject.toml
[project]
name = "e2b-research-agent"
version = "0.1.0"
description = "Research Agent using LangGraph"
authors = [
    {name = "User", email = "user@example.com"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=0.3.9",
    "langchain_openai>=0.2.11",
    "langchain_anthropic>=0.3.10",
    "langchain_groq>=0.0.3",
    "groq>=0.4.2",
    "openai>=1.6.0",
    "anthropic>=0.49.0",
    "e2b_code_interpreter>=1.1.1",
    "tenacity>=8.5.0",
    "pandas==2.2.1",
    "matplotlib==3.8.3",
    "seaborn==0.13.2",
    "scikit-learn==1.6.1",
    "langgraph>=0.3.16",
    "tavily-python>=0.5.0",
    "langchain-core>=0.3.45",
    "requests>=2.32.0",
    "typing-extensions>=4.0.0",
    "pydantic>=2.5.2",
    "tiktoken>=0.5.1",
    "mcp>=1.4.1",
    "langchain-mcp-adapters>=0.0.6"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta" 